重构 SFTP 模块边界
